Updated README.md V3
Finished describing the entities. I still have some small patches to perform on the README.md file such as uploading the UML diagram of the class hierarchy or adding/removing some business logic. This commit was performed earlier today at 15:04:22 but I had to edit it and push it again.