Started Working On The "Add Person" Option
Currently, I've only implemented the menu part, but I'm trying to redesign the code such that I have an interface and a service class strictly for the Person base class. The Person POJO class's service class will be called from the general "Service Menu" class whenever the "A" option will be selected. Furthermore, I will be trying to make the Person's service class call the Customer's and the Employee's service classes in order to better differntiate between which type of Person extension we're adding to our general Person ArrayList.