Finished Adding Employee + Added "input.txt" File
I still have to implement adding Customer as well. Furthermore, I have to focus at fixing the problem with temporary objects taking reserving ID numbers. For example, when I add Employees, their ID number starts with 3, then 6 and so on, this happens because creating an Employee implies creating a temporary Person and Employee object, which take up the ID numbers 1 and 2 before the actual Employee object that we want to create takes the ID number 3.