Bug Fix: ID Numbers
The last commit was slightly wrong: the description was correct but it didn't correspond to the code. In the code we were actually creating three temporary objects(2 Persons and 1 Employee). The first Person was created in PersonService.readPerson() on line 31, the second Person was created ACCIDENTALLY in EmployeeService.readEmployee() on line 25, the third person was actually an Employee created in EmployeeService.readEmployee() on line 53. The description of the last commit was correct: we really needed a temporary Person and one temporary Employee in order to create a correct Employee object. The mistake was that we created a second temporary Person object when we were trying to pull the returned temporary Person object from PersonService.readPerson(), this happened because we were actually creating a new copy of that object(using new Person()) instead of just referencing it. The fix was that we simply changed the line 25 of EmployeeService.readEmployee() such that it only references the returned object, it doesn't copy it. Thus, we will increase the ID once for the first temporary Person object and then immediately decrease it, we reference that object and create a new temporary Employee but this time we don't decrease the ID counter. I'm still undecided if we should add the reference of the temporary Employee created in EmployeeService.readEmployee() to the personList or if we should create a copy of the EmployeeService.readEmployee() returned object in PersonService.addPerson(), instead of using the reference.