Pull Request Fix: Fixed ID Assignment Issue On Derived Class Instance Creation
The problem was with the "removeTemporaryObject()" methods, since in the last commit we've transformed the base classes to abstract classes, there were no more temporary base class objects created(abstract classes donot allow instances to be created). In this commit we've removed that method entirely since we're not going to be needing it anymore.