Pull Request Fix: Fixed Input Bugs
Starting from this commit, the application will have no problem with intaking strings when it requires numbers as input. The only problem that remains is when an invalid date is typed. My guess is that I can solve this either by using regular expressions or by implmenting a Date class.