Added Generic CSV Reader + Improved Audit Logging
I still have to do the automatic data loader for when the application gets executed. The CSV Reader will be reading the corresponding CSV file and will return a collection in the form of a matrix with the contents of that file, ignoring the header line, obviously. The audit got improved by the fact that now it gets created automatically if it didn't exist before the application's execution.